Sorry for the mess.
read_and_display now displays the message if we couldn't parse it.
Parse way more messages, and add handlers in bot.c for them.
Add two useless functions of the bot:
-netcat to it on port 4242 will answer 0 or 1 depending on,
if you've got a highlight since last time a quassel client was started.
-Saying !lastlog something will search the last 60 lines of log of #quassel
And tell them to the user who asked it